Move sqlite activity
